+++ src/map_service/path/finder/finder.h
@@ -11,7 +11,7 @@ class PathFinder
     PathFinder() = default;
 
     std::vector<size_t> FindPath(const size_t &from, const size_t &to, const size_t &size,
-                                 const std::vector<std::vector<int>> &AdjacencyTable);
+                                 const std::vector<std::vector<double>> &AdjacencyTable);
 
     ~PathFinder() = default;
 };
